refactor(SocialButtons): changes files to default pattern of project
