--- src/components/Footer/SocialShareButtons/index.ts
@@ -1,1 +1,0 @@
-export { SocialShareButtons } from './SocialShareButtons';
